Use non-blocking connect for TcpStream.
Instead of spawning a background thread which is unaware of any timeouts
but continues to run until the TCP stack decides that the remote is not
reachable we use mio's non-blocking connect.

mio's `TcpStream::connect` returns immediately but the actual connection
is usually just in progress and we have to be sure the socket is
writeable before we can consider the connection as established.
Therefore we register to be notified for write readiness. To ensure we
get notified we create the watcher as level-triggered. Otherwise the
connect operation might have finished before we even registered
ourselves and an edge-triggered registration will not wake us up.